+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### Version 5.4.0 (xxth June 2025)
+#### Added
+- Added support for configuring store information via the `AdjustStoreInfo` object. You can now specify the store name and store app ID by assigning the `storeInfo` member of your `AdjustConfig` instance. This enables the SDK to record the intended app store source during initialization. For more details about this feature, refer to the [official documentation](https://dev.adjust.com/en/sdk/corona/setup/store-type).
+- Added ability to initialize the SDK for the first session in delayed mode. You can start the SDK in the delayed mode by setting the `isFirstSessionDelayEnabled` member on your `AdjustConfig` instance to `true`. To end the delay, make sure to call `endFirstSessionDelay` method of `Adjust` instance. For more details about this feature, refer to the [official documentation](https://dev.adjust.com/en/sdk/corona/features/first-session-delay).
+- Added ability to send organic search referrer together with deep link. You can send it via `referrer` member of the `AdjustDeeplink` instance. For more details about this feature, refer to the [official documentation](https://dev.adjust.com/en/sdk/corona/features/deep-links#handling-deeplinks-with-referrer).
+- Added ability to disable SDK's interaction with `AppTrackingTransparency.framework` API. You can disable it by setting the `isAppTrackingTransparencyUsageEnabled` member on your `AdjustConfig` instance to `false`. For more details about this feature, refer to the [official documentation](https://dev.adjust.com/en/sdk/corona/features/att#disable-att-framework).
+- Added `jsonResponse` field (JSON string) to `AdjustAttribution` where every key-value pair sent by the backend as part of the attribution response can be found.
+
+#### Native SDKs
+- [iOS@v5.4.0][ios_sdk_v5.4.0]
+- [Android@v5.4.0][android_sdk_v5.4.0]
+
+---
 ### Version 5.0.1 (14th February 2025)
 #### Added
 - Added sending of the additional parameter to improve troubleshooting of `SKAdNetwork` related issues.